Develop

See merge request module-partners/ebay2!6

+++ classes/tabs/EbayFormShippingTab.php
@@ -18,9 +18,9 @@
  * versions in the future. If you wish to customize PrestaShop for your
  * needs please refer to http://www.prestashop.com for more information.
  *
- *  @author    PrestaShop SA <contact@prestashop.com>
- *  @copyright 2007-2019 PrestaShop SA
- *  @license   http://opensource.org/licenses/afl-3.0.php  Academic Free License (AFL 3.0)
+ * @author 202-ecommerce <tech@202-ecommerce.com>
+ * @copyright Copyright (c) 2017-2020 202-ecommerce
+ * @license Commercial license
  *  International Registered Trademark & Property of PrestaShop SA
  */
 
@@ -139,7 +139,7 @@ class EbayFormShippingTab extends EbayTab
                     if ($ship_cost == false) {
                         $ship_cost = 0;
                     }
-                    if (isset($extra_fees) && (int)$ship_cost < $extra_fees[$key]) {
+                    if (isset($extra_fees) && (float)$ship_cost < $extra_fees[$key]) {
                         return false;
                     }
